requirements.txt
python-pptx
deepl
python-dotenv
